app-deploy-ops/requirements.txt
paramiko>=3.0
pyyaml>=6.0
jinja2>=3.1